Move depth_for_scale into index::builder, delete healpix module
The only custom HEALPix function left was depth_for_scale, used
only by the index builder and CLI. Move it there and remove the
now-empty healpix module entirely.
